+++ decode.h
@@ -1,8 +1,8 @@
 #ifndef DECODE_HEADER
 #define DECODE_HEADER
 
-char decode(unsigned int encoded_char);
-void verbose_decode(unsigned int encoded_char, char *buffer);
+char decode(unsigned long encoded_char);
+void verbose_decode(unsigned long encoded_char, char *buffer);
 
 #endif
